requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
comm==0.2.2
coverage==7.6.10
cryptography==44.0.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
fqdn==1.5.1
fsspec==2024.3.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.1.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kaitaistruct==0.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.22
Pygments==2.17.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.28.1
selenium-wire==5.1.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
torch==2.2.2
tornado==6.4
traitlets==5.14.2
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
uri-template==1.3.0
urllib3==1.26.12
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wsproto==1.2.0
zstandard==0.23.0
